RM#23390: add missing definition of action-invoice-payment-method-check-and-save and relocate call in the correct controller
